+++ gar/filters.py
@@ -9,6 +9,9 @@ from . import models as gar_models
 
 class GarFilter(DateTimeFilterMixin, FilterSet):
     bro_id__icontains = filters.CharFilter(field_name="bro_id", lookup_expr="icontains")
+    gmw_bro_id__icontains = filters.CharFilter(
+        field_name="gmw_bro_id", lookup_expr="icontains"
+    )
 
     class Meta:
         model = gar_models.GAR
@@ -18,3 +21,27 @@ class GarFilter(DateTimeFilterMixin, FilterSet):
                 "filter_class": CharFilter,
             },
         }
+
+
+class FieldMeasurementFilter(DateTimeFilterMixin, FilterSet):
+    class Meta:
+        model = gar_models.FieldMeasurement
+        fields = "__all__"
+
+
+class LaboratoryResearchFilter(DateTimeFilterMixin, FilterSet):
+    class Meta:
+        model = gar_models.LaboratoryResearch
+        fields = "__all__"
+
+
+class AnalysisProcessFilter(DateTimeFilterMixin, FilterSet):
+    class Meta:
+        model = gar_models.AnalysisProcess
+        fields = "__all__"
+
+
+class AnalysisFilter(DateTimeFilterMixin, FilterSet):
+    class Meta:
+        model = gar_models.Analysis
+        fields = "__all__"