+++ CHANGES.md
@@ -3,7 +3,10 @@
 ## 1.69 (unreleased)
 
 
-- Nothing changed yet.
+-   Enhancement [GAR]: Add all GAR models, including API pages, serializers and filters
+-   Enhancement [FRD]: Add all FRD models, including API pages, serializers and filters
+-   Enhancement [GAR]: Improve GAR-import to include all data, instead of just the main information of GAR
+-   Enhancement [FRD]: Improve FRD-import to include all data, instead of just the main dossier.
 
 
 ## 1.68 (2026-04-10)